refactor: Remove unused orderNumber field
The `orderNumber` field in the `Order` interface and its usage throughout the application are redundant. The unique `id` of the order, sliced to its last 6 characters, provides sufficient identification and is already used in relevant places. This change removes the `orderNumber` field and updates all instances where it was previously referenced.